Refactor menu buttons
