+++ packages/react-native-eishis-components-example/src/App.js
@@ -1,59 +1,10 @@
-import React from 'react';
-import { Text, TouchableOpacity, StyleSheet } from 'react-native';
-import {
-  createDrawerNavigator,
-  createStackNavigator,
-  DrawerActions,
-} from 'react-navigation';
-import Buttons from './screens/Buttons';
-import Checkboxes from './screens/Checkboxes';
-
-const styles = StyleSheet.create({
-  menuButton: {
-    height: '100%',
-    width: 60,
-    alignItems: 'center',
-    justifyContent: 'center',
-  }
-});
-
-const ButtonsStack = createStackNavigator({
-  Buttons: {
-    screen: Buttons,
-    navigationOptions: ({ navigation }) => ({
-      title: 'Buttons',
-      headerLeft: (
-        <TouchableOpacity
-          style={styles.menuButton}
-          onPress={() => navigation.dispatch(DrawerActions.openDrawer())}
-        >
-          <Text>Menu</Text>
-        </TouchableOpacity>
-      ),
-    }),
-  },
-});
-
-const CheckboxesStack = createStackNavigator({
-  Checkboxes: {
-    screen: Checkboxes,
-    navigationOptions: ({ navigation }) => ({
-      title: 'Checkboxes',
-      headerLeft: (
-        <TouchableOpacity
-          style={styles.menuButton}
-          onPress={() => navigation.dispatch(DrawerActions.openDrawer())}
-        >
-          <Text>Menu</Text>
-        </TouchableOpacity>
-      ),
-    }),
-  },
-});
+import { createDrawerNavigator } from 'react-navigation';
+import ButtonsScreen from './screens/Buttons';
+import CheckboxesScreen from './screens/Checkboxes';
 
 const AppNavigator = createDrawerNavigator({
-  Buttons: { screen: ButtonsStack },
-  Checkbox: { screen: CheckboxesStack },
+  Button: { screen: ButtonsScreen },
+  Checkbox: { screen: CheckboxesScreen },
 })
 
 export default AppNavigator;
